fix: fixing cursor mess up
